start to implement input and output types for m,odel functions
