+++ butilib/__init__.py
@@ -1,3 +1,5 @@
-from .schema import Suit, OROS, BASTOS, COPAS, ESPADAS
+from .suit import Suit, OROS, BASTOS, COPAS, ESPADAS
 from .card import Card
 from .deck import Deck
+from .model import Model
+from .schema import CantarInput
