+++ butilib/card.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, conint
 
-from .schema import Suit
+from .suit import Suit
 
 class Card (BaseModel) :
     number: conint(ge=1, le=12)
