+++ butilib/model.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class Model (BaseModel) :
+    pass
