+++ butilib/suit.py
@@ -1,0 +1,11 @@
+from enum import Enum
+class Suit (Enum) :
+    OROS = "OROS"
+    BASTOS = "BASTOS"
+    ESPADAS = "ESPADAS"
+    COPAS = "COPAS"
+    
+OROS = Suit.OROS
+COPAS = Suit.COPAS
+ESPADAS = Suit.ESPADAS
+BASTOS = Suit.BASTOS
