+++ tests/test_play.py
@@ -1,0 +1,1 @@
+import pytest
