+++ tests/test_schema.py
@@ -1,13 +1,23 @@
-import butilib
+import pydantic
 import pytest
-from enum import Enum
+import butilib
 
-def test_butilib_has_suit_enum_with_values_oros_copas_espadas_bastos () :
-    assert issubclass(butilib.Suit, Enum)
-    assert butilib.Suit._member_names_ == ["OROS", "BASTOS", "ESPADAS", "COPAS"]
+def test_cantar_input_is_sublclass_of_pydantic_base_model () :
+    assert issubclass(butilib.CantarInput, pydantic.BaseModel)
     
-def test_butilib_allows_for_importing_individual_suit_values () :
-    assert butilib.OROS == butilib.Suit.OROS
-    assert butilib.COPAS == butilib.Suit.COPAS
-    assert butilib.ESPADAS == butilib.Suit.ESPADAS
-    assert butilib.BASTOS == butilib.Suit.BASTOS
+def test_cantar_input_has_cards_and_delegated_fields_and_expects_exactly_twelve_cards () :
+    deck = butilib.Deck.new()
+    
+    card_list = deck.pop_some(12)
+    cantar_input = butilib.CantarInput(cards=card_list, delegated=False)
+    assert isinstance(cantar_input, butilib.CantarInput)
+    assert cantar_input.cards == card_list
+    assert cantar_input.delegated == False
+    
+    card_list = deck.pop_some(6)
+    pytest.raises(pydantic.ValidationError, butilib.CantarInput, cards=card_list, delegated=True)
+    
+    card_list = deck.pop_some(14)
+    pytest.raises(pydantic.ValidationError, butilib.CantarInput, cards=card_list, delegated=True)
+    
+
